ticket:1206 - Handle axis iterator trimming expression
Summary: When try to calculate axis iterator from Coverage Constructor (over $t (0:20)) it will calculate it such as a geo-referenced axis with upperLimit - 1, due to CRS is not CRS:1. Then, set this case to CRS:1 and it will return the same axis domain interval as Grid interval case.

Test Plan: Run the test case wcps test with ticket1202

Reviewers: mdumitru, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D157